test: add deterministic test suite and CI
171 pytest tests replace six manual scripts that had to be run by hand and
needed a browser, a Reddit session and an Ollama daemon to say anything.

The default run is offline by construction: a conftest fixture replaces
socket.connect so any accidental network call fails the test rather than
quietly succeeding. External providers are stubbed at their boundary.

Coverage is aimed at the parts that break quietly - config parsing and
validation, caption timing and overlap, FFmpeg command construction, the state
store, metadata fallback, error paths and CLI exit codes. One integration test
renders a real MP4 and probes its geometry, codecs and duration; it skips
itself when FFmpeg is absent.

CI lints, byte-compiles, runs the suite on Linux, macOS and Windows across
Python 3.11 and 3.13, builds the wheel, and installs it into a clean
environment to prove the console script works.

+++ src/resumex/narration/silent.py
@@ -54,7 +54,7 @@ class SilentNarrator(Narrator):
 def _write_silence(destination: Path, duration: float) -> None:
     """Write a mono 16-bit WAV of the given length."""
     destination.parent.mkdir(parents=True, exist_ok=True)
-    frames = max(1, int(round(duration * SAMPLE_RATE)))
+    frames = max(1, round(duration * SAMPLE_RATE))
     with wave.open(str(destination), "wb") as handle:
         handle.setnchannels(1)
         handle.setsampwidth(SAMPLE_WIDTH)
